+++ metadata.rb
@@ -9,12 +9,10 @@ version '2.0.0'
 depends 'nginx'
 depends 'python'
 depends 'runit'
-depends 'supervisor'
 %w(ubuntu centos).each do |platform|
   supports platform
 end
 
 recipe 'devpi::server', 'Installs the devpi-server package.'
-recipe 'devpi::supervisor', 'Manage the devpi-server daemon with supervisord.'
 recipe 'devpi::nginx', 'Configure nginx as an HTTP front-end.'
 recipe 'devpi::client', 'Install the devpi-client package.'