+++ test/cookbooks/supervisor-daemon/metadata.rb
@@ -2,9 +2,9 @@ name 'supervisor-daemon'
 maintainer 'Dave Shawley'
 maintainer_email 'daveshawley@gmail.com'
 license 'Apache 2.0'
-description 'Installs supervisor daemon on RHEL.'
+description 'Installs supervisor daemon.'
 long_description('This recipe installs the init-based startup ' \
   'files for supervisor.')
-version '0.1.0'
+version '0.2.0'
 
 depends 'supervisor'