+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This file is used to list changes made in each version of the devpi cookbook.
 ## 2.0.0
 - Fully embrace ChefDK for development.
 - Switch attribute access from symbols to strings.
+- Remove supervisor recipe.
 
 ## 1.1.0 (Never Released)
 - Add the *client* recipe that installs the `devpi` command line utility.